requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
google-genai==0.3.0
websockets==13.1
